requirements.txt
hidapi>=0.14.0
pyusb>=1.2.1
customtkinter>=5.2.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.12.0
